requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.5
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
filelock==3.0.12
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.3
